Former-commit-id: 5fc8862463210c2bea73195d363da1d0089f57a2

+++ app/Http/Controllers/UsersController.php
@@ -179,6 +179,10 @@ class UsersController extends Controller
 
     $user->update($requestForm);
 
+    if($request->ajax()) {
+      return response()->json('User updated successfully!');
+    }
+
     $request->session()->flash('message.nivel', 'success');
     $request->session()->flash('message.title', 'Well done!');
     $request->session()->flash('message.content', 'You upgrade has been success ');
@@ -192,17 +196,22 @@ class UsersController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+
   public function destroy($id)
   {
     $user = User::find($id);
     $user->delete();
-    
+
     //Crisp Delete 
     $CrispClient = new EventCrisp();
-    $people = $CrispClient->deleteProfile($user->email);
-    
+    $exist =  $CrispClient->checkIfExist($user->email);
+    if($exist == 'true'){//Eliminamos el perfil
+      $people = $CrispClient->deleteProfile($user->email);
+    }
     return $user;
   }
+
+
   public function destroyUser(Request $request,$id)
   {
     if($request->user_id){
